Move errors to new tangle pkg
