+++ pkg/dag/cone_root_indexes.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gohornet/hornet/pkg/model/hornet"
 	"github.com/gohornet/hornet/pkg/model/milestone"
 	"github.com/gohornet/hornet/pkg/model/storage"
-	"github.com/gohornet/hornet/pkg/model/tangle"
+	"github.com/gohornet/hornet/pkg/tangle"
 )
 
 // updateOutdatedConeRootIndexes updates the cone root indexes of the given messages.
